feat: wire Phase 5 commands in main.rs + mod.rs
Register ecosystem, logs, metrics_cmd, monitor, prime, replay, watch_cmd
in src/commands/mod.rs. Replace not_yet_implemented stubs for Prime,
Monitor, Logs, Watch, Ecosystem, Replay, Metrics with proper args structs
and execute calls. Extend WatchArgs with Stop/Status subcommands to wire
execute_stop/execute_status. Fix pre-existing type errors in logs.rs,
replay.rs, ecosystem.rs, and metrics_cmd.rs exposed by adding modules.

+++ src/commands/metrics_cmd.rs
@@ -193,7 +193,7 @@ mod tests {
         capability: &str,
         duration_ms: i64,
         completed: bool,
-        exit_code: Option<i32>,
+        exit_code: Option<i64>,
     ) -> SessionMetrics {
         SessionMetrics {
             agent_name: agent.to_string(),
